CORE: implemented fan shaped attacks. fixed a bug in AoE targeting. Poison breath no longer lasts forever. Added some food animations

+++ src/map/ai/ai_mob_dummy.h
@@ -63,7 +63,7 @@ protected:
   uint32 m_LastRangedTime;
   uint32 m_LastWaitTime;
   uint32 m_WaitTime;
-  uint16 m_skillTP;
+  float m_skillTP;
 
 	void ActionRoaming();
 	void ActionEngage();
